test: add a basic test
