+++ packages/postman-to-openapi/env.d.ts
@@ -1,0 +1,4 @@
+declare module '*?raw' {
+  const content: string
+  export default content
+}
